Some textures, and a big light
Added a couple blocks and some ground and threw some concrete textures
on them. Got a nifty skybox for the nightime and made a big tower that
shines a light down. I was trying to get the tower to randomly scan
around but the script is super buggy. Also I was playing with the
ambient light and stuff and the low lighting makes the concrete textures
look really oily, but that can be fixed.